requirements.txt
anthropic
notion-client
python-dotenv
requests
beautifulsoup4
